scheduling-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
google-cloud-aiplatform==1.43.0
google-auth==2.25.2
redis==5.0.1
pydantic==2.5.0
python-dateutil==2.8.2
ortools==9.8.3296
numpy==1.24.3
httpx==0.25.2
python-dotenv==1.0.0
structlog==23.2.0
